cli: add params for creating package with properties
To be able to create non Customer namespace packages I need to be able
to set SoftwareComponent, ApplicationComponent and TransportLayer.

Unfortunately, I also need to pass Correction Number but this
information is required by other ADT objects and hence I want to
introduce an approach to avoid the need to add a parameter for
Correction number to all commands manually.