adt: add support for changing of main code of classes
The method is almost the same as the program's one. The only difference is that
there are different headers - the class' one adds Accept.

I believe the method change_text should be factored out of the ADT object and
moved into a strategy/visitor class which would be able to deal with all object
types and could provide enough room for dealing with ADT implementation
differences.