adt: add fetch to ADT object
Retrieves ADT XML data, deserializes it and stores the values in object
properties

I had to add setter for the property active because the Marshal tries to
set all discovered attributes and this information must be always
downloaded. Hence I didn't lose time by adding the possibility to ignore
certain attributes in deserialization.