cli: make checkout create abapgit like directories
Create the root directory of the checked-out package name and create
the file .abapgit.xml in there.

This is a quick hack! I have a plan to add a module with ABAP types,
factory methods with the needed defaults and serialized/deserializer
for XML.

Thank you for the initial testing and
feedback, Christoph Pohl <christoph.pohl@sap.com>!