xml: enhance marshalling to create new objects
If you need to create an object with a property which should return None
by default but you also need to create the child object during
de-serialization, set the factory parameter of the decorator xml_element
and make sure you configure setter for the xml_element property. The
class Marshal will use the attribute factory to create an new instance
and set it back through the standard setter.

The factory method is called even when there is no standard setter but
this setup is not much helpful.