activation: activate found children objects
For example ABAP Class methods are ABAP Class children which are not
automatically activated whit activation of the parent class but
the first activation call referencing only the class returns a list of
inactive objects instead of activating the class and one must send back
the list to confirm the activation - it is driven by the HTTP parameter
preauditRequested.

I enhanced the marshalling framework to be able to serialize List and
to work also with "ADTObject"-like objects because WB classes should
not inherit from ADTObject but should have ADTObject traits. For example
not all serialized objects have adtcore:type attribute or some objects
can be from adtcore namespace (i.e IOCObjecReferences).

Marshalling no longer ignores the attributes named with '_' (single
underscore prefix) because Python internals starts with '__' and I want
to add a private helper properties creating a new item for a private
list.

Closes #7