adt: define __str__ for adtobjects
I want to print out a written object name but in the case of ABAP Class
we might be writing an include.

I want to print out an activated object but in the case of ABAP Class
we always activate the class even when writing an include.

Hence, for the former, we need __str__, and for the latter, we just use
the property name.