build: add coverage
The coverage report command does not rely on the .coveragerc run source
option because it does not include the files that has no coverage at
all.

It is clear that I do not want to run coverage for every 'make test'
because that seems to be superfluous as I run tests very often and I
think I do not need to see the coverage for each run (that should be
part of git commit hook and git pull request voter).