adt: annotation XML Element & Attribute Descriptor
The descriptor has 1 instance per a class so the descriptor must create
a new class attribute upon on the first call of the setter.

Avoid the need to always declared getter and setter with XML
annotations when there is no logic in these methods.

ADT uses plethora of types, so writing corresponding classes is
extremely annoying.