adt: fix deserialization of Text Lists
The problem was that Marshall was using the target property as temporary
storage - first it put the empty string, then it appended chunk - and
fall all the operations it was using the target property.

The lists are hacky and the setter appends a new value, so the above
algorithm causes that the empty string was added to the resulting list.

Also the operation of appending chunk produced invalid values because
the getter returns the list and not the current value. Therefore the
concatenation resulted into a list joined with a string (the chunk).