platform: add ABAP object XML deserializer
We will need this for abapGit configuration file and
also for parsing errors returned by ADT.

I hope the solution is as straightforward as possible.
We needed to implement a stack because we have to deal with nested
structures. It could be achieved by recursion or by using queue.

However I decided to use write objects because we have to deal with
3 kinds of structures:
    - literal
    - structure
    - table (which can contain literals, structures or tables)