adt: get required FM MIME from Function Groups conf
Function Modules are dependant on Function Groups and ADT reports
the required version for Function Groups only, hence sapcli uses
a wrong version for Function Modules (it uses always the latest V3 while
older system requires V2).

This commit changes the behavior of Function Modules in the way that the
required mime version is retrieved for Function Groups.