gcts: suppor no-fail-exist for cloned repos too
A cloned repository cannot be cloned again and gCTS dies with 500.
I did not use the approach try & check because here I think it's
safer to check the state - I am not sure what could possibly happen in
ABAP back-end if we try to clone a cloned repo in future.