aunit: fix error for missing executionTime in ADT response
I don't know why ADT sometimes leaves out executionTime but I should
try to find out once to make sure the issues does not re-occur.

However, for the time being, this simple fix is enough.

If you are reading this because you are not sure how to deal with
my commit, you are free revert and do better.

The exception string:
  TypeError: float() argument must be a string or a number, not 'NoneType'