tree-wide: add create, read, change user operations
The CLI operations are very limited because my main goal was to
implement the operation change in order to be able to change users
password.

The other code which is no directly enabling the reset password
operation is there just to let me test the reset password operation
working.

Therefore, any enhancements to those other operations are more than
welcome!