cli: Enhance error output of exception type
The motivation for this is case, where remote server provides wrong or
misleading error description, which doesn't explain what really happened
and what is the root cause. Sapcli can handle such cases, but it stores
this information only as proper exception type. If the exception type
isn't printed out, this information is lost.

The example is creation of the function group via ADT, where creation
operation fails in case the function group already exists. ADT provides
completely wrong error message as a response.