adt: connection refreshes CSRF token upon 403
CSRF token gets invalidated after a long period of inactivity.

I don't want to fetch the token before each request to make it more
efficient. The pattern is clear - if you get 403 and it's not "Fetch
the initial token" we can expect we ran into the issue with invalidated
token.