adt: replace Windows Endings with Unix in object sources
The weirdest thing is that ADT returns XML with the first line ended by
Unix line ending and the rest of the lines with Windows line endings.

It confuses text editors like vim which displays ^M for the superfluous
'\r'.

This patch unifies the line endings to Unix ones but if necessary,
it can be made configurable.