gcts: print repolist as table
The table output is the key point but the alignment of the values is
the goal of the patch. Not aligned values are hard to read with more
than 2 repos.

I had to create a helper class because Pylint was complaining
about too many local variables. I created the class somehow
generic, so it could be reused somewhere else.

TODO: move the helper class to a common module and add more tests.