adt: make the Object references class de-serializable
I need it for the ADT search functionality.

XmlListNodeProperty has strange behavior I did not document!

If you attempt to set such a property, the passed value is appended to
the list behind the property -> the value will not replace the current
property's value.

I was thinking about changing the implementation to recognize an
instance of list but I would not be able to build list of lists then.