doc: update strust command documentation
--- v2
Incorporated review comments: Added explanation of Storages.
In addition, I fixed the typo in the name of the Client Standard
storage.

--- v3
Incorporated review comments: Storage client_standart deprecated
For the Deprecation warning the built-in warnings module is used. By
default the DeprecationWarnings are disabled; thus I enabled it in the
main function.