gcts: add support of URL in PACKAGE argument
The PACKAGE argument still supports the repository name, but if the
argument starts with 'http://' or 'https://' it is meant as repository
URL.

If the command is called with URL, it will fail either when there are
no repositories with this URL or when there are multiple repositories
with the same URL.

--- v2
Incorrporated review comments: Fixed the exception that will be raise in
case of URL problem. Created test for 'delete()' when 'repo' is
provided.