cli: make TableWriter more versatile
--- v2 (palubaj)
I have changed the default value of column's parameter `default` to
sentinel object. It is necessary because object that is passed as
`data` could have defined an attribute to `None`, in which case it
ignores the `default` parameter of column and uses `None` as value.

--- v3 (jfilak)

I did a small correction to make sure it's perfectly OK if data contains
None. We do want to report error only if default is not set and key/attr
is missing.
I also decided to call str(val) only if FORMATTER is not set, to allow
TableWriter users to deal with None themselves without the need to do
`if val == 'None':`.