cli: allow specification of wrappers for commands
Current command's decorators did not allow the conventional use of
exception handler as additional decorator for commands. Mainly because
the function name is used as identificator by CommandGroup methods.

Because of that I have added the option to specify class attribute
'wrapper' for CommandGroup which wraps the commands function while it is
being processed.