pylint: fix unnecessary else after return or raise
The new version of pylint throws errors if there is 'else' statement
after return of raise, as it can be ommited.

In addition, pylint issued a warning that from the version 3.0 it will
no longer support exception names in overgeneral-exceptions without
module name. Because of that the '.pylintrc' was updated.