tree-wide: update JUnit AUnit output for Skipped tests
I was not sure how to convert ABAP ADT AUnit XML into JUnit.

I have decided to consider skipped only the tests with the kind ABORTION
and the severity TOLERABLE - I consulted ABAP code of
cl_abap_unit_assert=>skip( ) which sets the attributes this way.

To be defensive: I would rather receive a complain that we do not
recognize skipped test than that we missed a real error and incorrectly
marked it as skipped.

Jenkins output:

    DO_THE_FAIL – ZCL_THEKING_MANUAL_HARDCORE=>LTCL_TEST

    Standard Output:
    Include: <ZCL_THEKING_MANUAL_HARDCORE===CCAU> Line: <19> (DO_THE_FAIL)

    Standard Error:
    Test execution skipped due to missing prerequisites
    Test 'LTCL_TEST->DO_THE_FAIL' in Main Program 'ZCL_THEKING_MANUAL_HARDCORE===CP'.

Closes #101