abap: implement parsing of simple xml objects
Examples of simple xml objects:
(Not part of any structure)
<asx:values>
  <SIMPLE_OBJECT>value</SIMPLE_OBJECT>
</asx:values>

(In Internal table)
<asx:values>
  <INCLUDES>
    <OBJ>first include</OBJ>
    <OBJ>second include</OBJ
  </INCLUDES>
</asx:value>>

Notice that the second case does not contain '<item>' element. This
change is necessary for parsing XML files, which define function groups.
Due to this changes the ABAP writers were simplified and do not require
start() and do_end() functions.