checkin: format the abapgit source file
Function modules can be in abapgit format, meaning that the parameters
are specified in special comment block. I have create a function which
parses parameters from this comment block and writes them in
ADT-readable format.

Note, for TABLES parameters the keyword `STRUCTURE` is used instead of
`TYPE`. I have tested it in SAP GUI. Assigning typing `TYPE` for a
TABLES parameter is represented as `STRUCTURE` in source code editor.
However, in ADT editor `TYPE` is used, thus `STRUCTURE` needs to be
converted to `TYPE`.

Original source code:
```
    FUNCTION ztest_function.
    *"----------------------------------------------------------------------
    *"*"Local Interface:
    *"  IMPORTING
    *"     VALUE(IV_PARAM1) TYPE  STRING
    *"  EXPORTING
    *"     VALUE(EV_PARAM2) TYPE  STRING
    *"  TABLES
    *"     ET_PARAM3 STRUCTURE  BAPIRET2
    *"----------------------------------------------------------------------
    ...
    ENDFUNCTION.
```
Formatted source code:
```
    FUNCTION ztest_function
    IMPORTING
    VALUE(IV_PARAM1) TYPE STRING
    EXPORTING
    VALUE(EV_PARAM2) TYPE STRING
    TABLES
    ET_PARAM3 TYPE  BAPIRET2
    .
    ...
    ENDFUNCTION.
```
The indentation of the formatted source code is fixed upon write to the
system.

---v2
Fixed parameter type parsing. Added tests.