cli: implement function group checkout
One thing missing is the short text of parameters. The short text should
be part of <RSFDO> tag (which is part of <DOCUMENATION>), but it
requires reading table `FUNCT`. I left it to a second iteration as it
needs a bit of rework to get RFC connection in addition to the ADT
connection.