gcts: enhance sugar context managers to use share logger
We plan to enhance the sugar operation clone with the options like
no-import or buffer-only. Those operation temporarily changes gCTS
configuration and for that activity we use context managers with
loggers. To avoid the need pass new Sugar Operation logger instance for
each sugar operation context manager we need to be able to pass only 1
instance of logger which would be able to handle updates from many
context managers.

There are two problems which we need to solve:
- the observer must be able to clear the recover after the configuration
  change is restored
- since the "print recover message" logic is getting more complicated we
  must enhance the context manager to trigger it on its own