gcts: detect existing report repositories by another user
For the time being, we expect repositories get names generated from the
repository URL and thus if a repository of a certain name exists, we
expect it is the same ABAP code and thus it does not matter who created
the repo.

I have now realized that we should check that the repository which
already exists have the same URL as the one we wanted to create. This
check is important for automations where repo name can be force and not
generated from the URL (e.g. "ci-repo").